Remove Monitor header from Transactions
